init app with gRPC

+++ internal/models/rating_category.go
@@ -1,7 +1,7 @@
 package models
 
 type RatingCategory struct {
-	ID     int    `json:"id" db:"id"`
-	Name   string `json:"name" db:"name"`
-	Weight int    `json:"weight" db:"weight"`
+	ID     int     `json:"id" db:"id"`
+	Name   string  `json:"name" db:"name"`
+	Weight float64 `json:"weight" db:"weight"`
 }
